Server/Misc: Fix format warnings.

+++ src/server/game/Entities/Player/Player.cpp
@@ -6823,9 +6823,7 @@ void Player::SendActionButtons(uint32 state) const
         buttonsTab[i].id = ab->GetAction();
         buttonsTab[i].unk = ab->GetHiType();
 
-        SF_LOG_DEBUG("network", "SendActionButtons actionId: %u lowtype: %u hiType: %u", ab->GetAction(), ab->GetType(), ab->GetHiType());
-
-
+        SF_LOG_DEBUG("network", "SendActionButtons actionId: %u lowtype: %u hiType: %u", ab->GetAction(), uint8(ab->GetType()), ab->GetHiType());
     }
 
     // Bits
